==> 00_usefulPlots.ipynb <==
## Imports

# Importing other packages
import timeit
import pandas as pd
import numpy as np
import neptune
import tempfile
import seaborn as sns
import matplotlib.pyplot as plt

import ipywidgets as widgets


import plotly.graph_objects as go
import plotly.io as pio

pio.renderers.default = "browser"

## Load data

df = pd.read_csv(f'./data/amazon_22.csv', parse_dates=['EffectiveDate'])

df['instanceFamily'].value_counts()

df['TermType'].value_counts()

### Find unique instance types & count

instanceTypes = df['instanceType'].value_counts()

# exclude some instance families
df = df[(df['instanceFamily'] != "FPGA Instances") & (
    df['instanceFamily'] != "Machine Learning ASIC Instances")]

### Choose instanceType

# choose instance specific instance type

# df = df[(df['instanceType'] == "r5d.8xlarge")]

### Count unique Locations

df['Location'].value_counts()

df['Location'].value_counts()

# # keep only the countries that have data in all years
# df = df[(df['Location'] != "Africa") & (df['Location'] != "MiddleEast") & (df['Location']
#                                                                            != "Canada") & (df['Location'] != "SouthAmerica") & (df['Location'] != "AWSGovCloud")]

# Convert EffectiveDate column to datetime format
df['EffectiveDate'] = pd.to_datetime(df['EffectiveDate'])


df = df[df['year'] != 2023]

df

df.info()

# Extract the year and quarter from the Date column
df['year'] = df['EffectiveDate'].dt.year
df['Quarter'] = df['EffectiveDate'].dt.quarter

# Combine year and quarter into a single column
df['YearQuarter'] = df['year'].astype(
    str) + '-Q' + df['Quarter'].astype(str)

# Sort the DataFrame by 'YearQuarter' column in ascending order
df = df.sort_values('YearQuarter')

# Reset the index
df = df.reset_index(drop=True)

# Convert 'YearQuarter' column to PeriodIndex with quarterly frequency
df['YearQuarter'] = pd.PeriodIndex(df['YearQuarter'], freq='Q')

# Convert Period values to strings
df['YearQuarter'] = df['YearQuarter'].astype(str)

## Export dataframe with quarters

# Group RHEL & SUSE to Linux and rename Red Hat Enterprise Linux with HA
df = df.replace({'RHEL': 'Linux', 'SUSE': 'Linux', "Red Hat Enterprise Linux with HA": "RHEL_HA"})

# replace USEast and USWest to US and EU to Europe
df = df.replace({'USEast': 'US', 'USWest': 'US', 'EU': 'Europe'})

df.rename(columns = {'Network Performance': 'networkPerformance'},inplace=True)

df.to_csv(f'./data/amazon_22_quarters.csv', index=False)

## Calculate statistical metrics 

# Calculate statistic metricsby group
stats = df.groupby(['YearQuarter', 'Location', 'instanceFamily'])['PricePerUnit'].agg(
    ['mean', 'std', 'median', 'max', 'min']).reset_index()
stats.rename(columns={'mean': 'mean_price', 'std': 'std_deviation',
             'max': 'maximum', 'min': 'minimum'}, inplace=True)


# # Calculate statistic metricsby group
# stats = df.groupby(['year', 'instanceFamily'])['PricePerUnit'].agg(
#     ['mean', 'std', 'median', 'max', 'min']).reset_index()
# stats.rename(columns={'mean': 'mean_price', 'std': 'std_deviation',
#              'max': 'maximum', 'min': 'minimum'}, inplace=True)

stats['YearQuarter'].value_counts()

# create a list for inst families
instFamilies = df['instanceFamily'].unique()
instFamilies


# # Loop through each instanceFamily and save the corresponding subset as a new DataFrame
# for instance_family in instFamilies:
#     subset = df[df['instanceFamily'] == instance_family]
#     subset.to_csv(f'./data/{instance_family}_subset.csv', index=False)

# Plots

## Plot specific instance type (by user)

import plotly.express as px

# Take instanceType as input from the user
instance_type = input("Enter the instance type: ")

# Filter the data for the specified instance type
filtered_data = df[df['instanceType'] == instance_type]

# # Extract the year and quarter from the Date column
# filtered_data['Year'] = filtered_data['EffectiveDate'].dt.year
# filtered_data['Quarter'] = filtered_data['EffectiveDate'].dt.quarter

# # Combine year and quarter into a single column
# filtered_data['YearQuarter'] = filtered_data['Year'].astype(
#     str) + '-Q' + filtered_data['Quarter'].astype(str)

# Calculate statistical metrics by year and quarter
stats_filtered = filtered_data.groupby('YearQuarter')['PricePerUnit'].agg(
    ['mean', 'std', 'median', 'max', 'min']).reset_index()
stats_filtered.rename(columns={'mean': 'mean_price', 'std': 'std_deviation',
                               'max': 'maximum', 'min': 'minimum'}, inplace=True)

# Create the line plot
fig = px.line(stats_filtered, x='YearQuarter', y='mean_price',
              template='presentation',
              title=f'Evolution of median price for {instance_type}')

fig.update_traces(mode='lines+markers')

fig.update_layout(
    xaxis=dict(title='Year and Quarter',
               tickfont=dict(size=12, family='Arial')),
    yaxis=dict(title='Price per Unit', tickfont=dict(size=12, family='Arial')),
    legend=dict(orientation='h', yanchor='bottom',
                y=1.02, xanchor='right', x=1),
    title=dict(font=dict(size=14, family='Arial')),
    font=dict(size=12, family='Arial')
)

fig.show()

## Evolution of median for specific instance - all Locations (by user)

import pandas as pd
import plotly.express as px

instance_family = input("Enter the instance family: ")

filtered_data = df[df['instanceFamily'] == instance_family]

# Convert 'YearQuarter' column to PeriodIndex with quarterly frequency
filtered_data['YearQuarter'] = pd.PeriodIndex(filtered_data['YearQuarter'], freq='Q')

# Convert Period values to strings
filtered_data['YearQuarter'] = filtered_data['YearQuarter'].astype(str)

# Calculate statistical metrics by year and quarter
stats_filtered = filtered_data.groupby('YearQuarter')['PricePerUnit'].agg(
    ['mean', 'std', 'median', 'max', 'min']).reset_index()
stats_filtered.rename(columns={'mean': 'mean_price', 'std': 'std_deviation',
                               'max': 'maximum', 'min': 'minimum'}, inplace=True)

fig = px.line(stats_filtered, x='YearQuarter', y='median',
              template='presentation',
              title=f'Evolution of median price for {instance_family} Instance Family')

fig.update_traces(mode='lines+markers')

fig.update_layout(
    xaxis=dict(title='Year and Quarter',
               tickfont=dict(size=20, family='Arial')),
    yaxis=dict(title='Price per Unit', tickfont=dict(size=20, family='Arial')),
    legend=dict(orientation='h', yanchor='bottom',
                y=1.02, xanchor='right', x=1),
    title=dict(font=dict(size=24, family='Arial')),
    font=dict(size=18, family='Arial')
)

fig.show()


## Plot specific instance Family per Location

import plotly.express as px

# Calculate statistic metrics by group
stats = df.groupby(['YearQuarter', 'Location', 'instanceFamily'])['PricePerUnit'].agg(
    ['mean', 'std', 'median', 'max', 'min']).reset_index()
stats.rename(columns={'mean': 'mean_price', 'std': 'std_deviation',
                      'max': 'maximum', 'min': 'minimum'}, inplace=True)

# Convert 'YearQuarter' column to PeriodIndex with quarterly frequency
stats['YearQuarter'] = pd.PeriodIndex(stats['YearQuarter'], freq='Q')

# Convert Period values to strings
stats['YearQuarter'] = stats['YearQuarter'].astype(str)

compute_optimized_data = stats[stats['instanceFamily'] == 'Compute optimized']

# Fill missing data points with NaN values
locations = compute_optimized_data['Location'].unique()
quarters = compute_optimized_data['YearQuarter'].unique()
combined_index = pd.MultiIndex.from_product([locations, quarters], names=['Location', 'YearQuarter'])
compute_optimized_data = compute_optimized_data.set_index(['Location', 'YearQuarter']).reindex(combined_index).reset_index()

# Sort the DataFrame by 'YearQuarter' column
compute_optimized_data.sort_values(by='YearQuarter', inplace=True)

# Extract 'median' column as a Series
median_series = compute_optimized_data['median']

# Interpolate missing values in the Series
interpolated_median = median_series.interpolate()

# Update the original DataFrame with the interpolated values
compute_optimized_data['median'] = interpolated_median

# Filter out interpolated values before the first valid data point for each location
compute_optimized_data = compute_optimized_data.groupby('Location').apply(lambda x: x.loc[x['median'].first_valid_index():]).reset_index(drop=True)

# Create the line plot
fig = px.line(compute_optimized_data, x='YearQuarter', y='median', color='Location',
              template='presentation',
              title='Evolution of median price for Compute Optimized Instance Family')

fig.update_traces(mode='lines+markers')

fig.update_layout(
    xaxis=dict(title='Year', tickfont=dict(size=20, family='Arial')),
    yaxis=dict(title='Price per Unit', tickfont=dict(size=20, family='Arial')),
    legend=dict(orientation='h', yanchor='bottom',
                y=1.02, xanchor='right', x=1),
    title=dict(font=dict(size=24, family='Arial')),
    font=dict(size=18, family='Arial')
)

fig.show()


import plotly.express as px

compute_optimized_data = stats[stats['instanceFamily'] == 'Compute optimized']

fig = px.line(compute_optimized_data, x='YearQuarter', y=['median'],
              template='presentation',
              title='Evolution of median price for Compute Optimized Instance Family',
              hover_name='instanceFamily')

fig.update_traces(mode='lines+markers')

fig.update_layout(
    xaxis=dict(title='Year', tickfont=dict(size=20, family='Arial')),
    yaxis=dict(title='Price per Unit', tickfont=dict(size=20, family='Arial')),
    legend=dict(orientation='h', yanchor='bottom',
                y=1.02, xanchor='right', x=1),
    title=dict(font=dict(size=24, family='Arial')),
    font=dict(size=18, family='Arial')
)

fig.show()

fig = px.scatter(df, x=df.index, y=df['PricePerUnit'], template='presentation',
                 labels=dict(x='Data points', y='Price per unit'),
                 color='PricePerUnit', title='Density of Price per year', facet_col='year',
                 category_orders={'year': [2016, 2017, 2018, 2019, 2020, 2021, 2022]})
# Modify font properties
fig.update_layout(
    font=dict(
        family='Arial',  # Change to your desired font family
        size=16,
    )
)  # Change to your desired font size
# fig.write_html("results/stats/density.html")
fig.show()

import plotly.express as px

fig = px.scatter_3d(df, x='year', y='instanceFamily', z='PricePerUnit', template='presentation',
                    labels=dict(z='Price per Hour'),
                    color='instanceFamily', title='Density of Price per Year',
                    category_orders={'year': [2016, 2017, 2018, 2019, 2020, 2021, 2022]})

fig.update_layout(
    legend=dict(orientation='h', yanchor='bottom',
                y=-0.1, xanchor='left', x=0),
    font=dict(size=16, family='Arial'),
    scene=dict(
        xaxis=dict(title_font=dict(size=20, family='Arial')),
        yaxis=dict(title_font=dict(size=20, family='Arial')),
        zaxis=dict(title_font=dict(size=20, family='Arial')),
        xaxis_title='', yaxis_title='', zaxis_title='Price($) per Hour'
    ),
    title=dict(font=dict(size=24, family='Arial'))
)

fig.update_traces(marker=dict(size=8))

fig.show()

import plotly.express as px

fig = px.scatter(df, x='year', y='instanceFamily', color='PricePerUnit', template='presentation',
                 labels=dict(x='Year', y='instanceFamily',
                             color='Price per Hour'),
                 color_continuous_scale='Viridis', title='Density of Price per Year')

fig.update_layout(
    legend=dict(orientation='v', yanchor='bottom',
                y=0.5, xanchor='right', x=0),
    font=dict(size=16, family='Arial'),
    xaxis=dict(title_font=dict(size=20, family='Arial')),
    yaxis=dict(title_font=dict(size=20, family='Arial')),
    title=dict(font=dict(size=24, family='Arial'))
)

fig.update_traces(marker=dict(size=8))

fig.show()

import plotly.graph_objects as go

fig = go.Figure(data=go.Histogram2d(
    x=df['year'],
    y=df['instanceFamily'],
    colorscale='YlGnBu',
    zauto=True,
    nbinsx=len(df['year'].unique()),
    nbinsy=len(df['instanceFamily'].unique()),
))

fig.update_layout(
    title='Density of Offers per Year',
    xaxis=dict(title='Year'),
    yaxis=dict(title='Instance Family'),
    coloraxis_colorbar=dict(title='Density'),
)

fig.show()


# 'Viridis'
# 'Cividis'
# 'Hot'
# 'Jet'
# 'Rainbow'
# 'Electric'
# 'Blues'
# 'Greens'
# 'Reds'
# 'YlOrRd'
# 'YlGnBu'
