==> 04a_create_daset_for_pred.ipynb <==
import mlflow.sklearn
import pandas as pd
import plotly.graph_objects as go


import my_functions as mf

# Import data for prediction

data = pd.read_csv(f'./data/amazon_22_quarters.csv',
                   parse_dates=['EffectiveDate'])

data.info()

data = data[['PricePerUnit', 'LeaseContractLength', 'PurchaseOption', 'vCPU',
             'Memory', 'operatingSystem', 'networkPerformance', 'StorageSize',
             'DiskType', 'Quarter', 'instanceType', 'instanceFamily', 'YearQuarter', 'year']]

count_instTypes = data['instanceType'].value_counts()

grouped_data = data.groupby(['instanceType', 'instanceFamily'])

df_grouped = grouped_data.mean().reset_index()

# Take instanceType as input from the user

# instance_type = "g5.2xlarge"
# operating_system = 'Windows'
# purchase_option = "Partial Upfront"

instance_type = input("Enter the instance type: ")
operating_system = input("Enter the operating system: ")
purchase_option = input("Enter the purchase option: ")


# filter by instance type data
# filtered_data = data[(data['instanceType'] == instance_type)
#                      & (data['year'] > 2021)]

filtered_data = data[(data['instanceType'] == instance_type)
                     & (data['operatingSystem'] == operating_system)
                     & (data['PurchaseOption'] == purchase_option)
                     & (data['year'] > 2016)]

# Group the filtered data by YearQuarter and calculate the mean price
grouped_data = filtered_data.groupby('YearQuarter')['PricePerUnit', 'LeaseContractLength', 'PurchaseOption',
                                                    'vCPU', 'Memory', 'operatingSystem',
                                                    'networkPerformance', 'StorageSize',
                                                    'DiskType', 'Quarter', 'instanceType',
                                                    'instanceFamily', 'YearQuarter', 'year'].mean()

# Convert the grouped data to a DataFrame
grouped_data = grouped_data.reset_index()


grouped_data['year'] = grouped_data['year'].astype(int)

if 'PurchaseOption' not in grouped_data.columns:
    grouped_data.insert(3, 'PurchaseOption', purchase_option)

if 'operatingSystem' not in grouped_data.columns:
    grouped_data.insert(5, 'operatingSystem', operating_system)

DiskType = "SSD"
if 'DiskType' not in grouped_data.columns:
    grouped_data.insert(5, 'DiskType', DiskType)

# Create a new DataFrame with the same values as the previous row
new_rows = pd.concat([grouped_data.iloc[[-1]]] * 4, ignore_index=True)

# Set the YearQuarter values for the new rows
last_year_quarter = pd.Period(grouped_data['YearQuarter'].iloc[-1])
new_rows['YearQuarter'] = pd.period_range(
    start=last_year_quarter + 1, periods=4, freq='Q')

# Change the year in the new rows
new_rows['year'] = new_rows['year'] + 1


# Append the new rows to the original DataFrame
grouped_data = pd.concat([grouped_data, new_rows], ignore_index=True)

grouped_data.to_csv(
    f'./data/{instance_type}_{operating_system}_{purchase_option}.csv', index=False)
